warp2api-main/pyproject.toml
[project]
name = "warptestui"
version = "0.1.0"
description = "Test UI for Warp"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "httpx[http2]",
    "protobuf>=4.0.0",
    "grpcio-tools",
    "python-dotenv",
    "websockets>=15.0.1",
    "requests>=2.32.5",
    "openai>=1.106.0",
]

[project.scripts]
warp-server = "server:main"
warp-test = "main:main"

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/simple"
default = true
